Drop explicit dependency of binary package on postgresql
The schema can be installed in a remote database so there's no need
to have one locally.

Dependency on dbpatch and tableversion is made "recommended" because
it is only needed for the loader script and only for finding their
own loader scripts (currently in the same package as the PostgreSQL
side, unfortunately). For this reason specific versions are
recommended (those that do provide the loader scripts)

Closes #55