Drop confusing output from dbpatch loader preflight call
The call is made to verify capabilities of the script, so is
passed an arbitrary schema which we don't want to see in output.
A typo made this schema name surface to user, which was confusing.

Since I was at it, I changed name of fake schema to "fake" (from
"public") so it is clear we're not really doing that load

This is a backport of 8542dc2c04fc2255d07bedc2e7e5f2620ad6d8fb